Prefix product services with Product, move services under utils to correct place and ditch the intermittent / redundant wrapper class for product serializer
